requirements.txt
streamlit
google-genai
pandas
ics